+++ CHANGES.rst
@@ -7,6 +7,7 @@
 
 - Sideview feature-complete for multiple results and grids (#806, #811, #812, #670, #808, #789)
 - Result Manager: fixed bug when working dir contained a revision with number 0 (#822)
+- Graph tool: maptools are disabled when tabs are switched (#824)
 
 
 3.1.3 (2023-03-21)