+++ CHANGES.rst
@@ -11,6 +11,7 @@
 - Graph tool: Fixed bug where deactivated plots were added again when parameter/units or absoluteness was changed (#825)
 - Graph tool: default variable for pump should be 'discharge pump' (#819)
 - Removed wiki related files (wiki has been disabled)
+- Graph tool: pump can now only be added as line feature (no longer also as node feature) (#829)
 
 
 3.1.3 (2023-03-21)