+++ CHANGES.rst
@@ -13,6 +13,7 @@
 - Removed wiki related files (wiki has been disabled)
 - Graph tool: pump can now only be added as line feature (no longer also as node feature) (#829)
 - Graph tool: change "Nodes" to "Nodes & cells" (#818)
+- Initial work on Watershed tool (#668)
 
 
 3.1.3 (2023-03-21)